feat: Save burned info at Commit
