+++ ctrlers/vpower/ctrler_block.go
@@ -25,6 +25,8 @@ func (ctrler *VPowerCtrler) BeginBlock(bctx *ctrlertypes.BlockContext) ([]abcity
 				ctrler.logger.Error("Error when punishing",
 					"byzantine", types.Address(evi.Validator.Address),
 					"evidenceType", abcitypes.EvidenceType_name[int32(evi.Type)])
+			} else if xerr = bctx.SupplyHandler.Burn(bctx, ctrlertypes.PowerToAmount(slashed)); xerr != nil {
+				return nil, xerr
 			} else {
 				evts = append(evts, abcitypes.Event{
 					Type: "punishment.stake",
